+++ tests/testthat/test-08_throwback_thursdays.R
@@ -1,3 +1,9 @@
+test_that("tt_review()", {
+  expect_output(tt <- tt_review())
+  expect_s3_class(tt, "data.frame")
+  expect_named(tt, c("post_date", "nb_visits", "url"))
+})
+
 test_that("tt_posts()", {
   withr::local_envvar(list(CLIPR_ALLOW = TRUE))